🔨 refactor: 메서드 분리
